requirements.txt
aiohttp==3.13.2
discord.py==2.6.4
python-dotenv==1.2.1
